Ethics-Codes-of-Conduct-2
Il.Terry Cooper: We would need a principle of justice to show us what pattern of action would reflect justice as a value. A common form of the justice principle is “Treat equals equally and unequals unequally.” To summarize, values are general moral obligations while principles are the ethical conditions or behaviors we expect. Unfortunately, this can become confusing in everyday language. For instance, many times “core values” or “concrete values” are terms used instead of principles but they should be intended to inform principles.
For this reason it is not uncommon for codes to being with a value (integrity) and then make the value real in principle (Do not use your public office for private gain). Many modern public services initiate their code development through input from public servants. 
14 Jeremy Bentham, An Introduction to The Principles of Morals and Legislation (NY: Hafner Press, 1948), p. 2. [Original publication date is 1780).
15 Terry L. Cooper, The Responsible Administrator: An Approach to Ethics in the Administrative Role, 4th edition (San Francisco, Jossey-Bass Publishers, 1998), p.